+++ test/unit/no-validate.node.js
@@ -1,0 +1,62 @@
+import assert from 'assert';
+import * as util from '../../dist/lib/util';
+
+import Core from '../../dist/lib/core';
+Core.plugin(require('../../plugins/no-validate'));
+Core.plugin(require('pouchdb-adapter-memory'));
+
+const schema = {
+    title: 'human schema',
+    description: 'describes a human being',
+    version: 0,
+    disableKeyCompression: true,
+    type: 'object',
+    properties: {
+        passportId: {
+            type: 'string',
+            index: true
+        },
+        firstName: {
+            type: 'string'
+        },
+        lastName: {
+            type: 'string'
+        }
+    },
+    required: ['firstName', 'lastName']
+};
+
+describe('no-validate.node.js', () => {
+    it('should allow to insert everything', async () => {
+        const db = await Core.create({
+            name: util.randomCouchString(10),
+            adapter: 'memory'
+        });
+        const col = await db.collection({
+            name: 'humans',
+            schema
+        });
+        await col.insert({
+            foo: 'bar'
+        });
+        db.destroy();
+    });
+    it('should allow to save everything', async () => {
+        const db = await Core.create({
+            name: util.randomCouchString(10),
+            adapter: 'memory'
+        });
+        const col = await db.collection({
+            name: 'humans',
+            schema
+        });
+        await col.insert({
+            foo: 'bar'
+        });
+        const doc = await col.findOne().exec();
+        assert.equal(doc.get('foo'), 'bar');
+        doc.set('bar', 'foo');
+        await doc.save();
+        db.destroy();
+    });
+});